test(guard-run): move tests to top-level test/ mirroring src
Reviewer convention: test/ sits beside src/ at repo root, each test file
at the same relative path under test/ as its subject under src/. The four
guard-run suites move src/test/guard-run/*.test.ts -> test/guard-run/, and
their imports go up one more level (../../guard-run -> ../../src/guard-run).

vitest include now lists Vitest's default glob (so packages/** and
internal/** stay discovered) plus test/**/*.test.ts for the new tree. Only
this PR's own tests move; the repo-wide src/test/** convention is untouched.
